Clearly separate imported content schema from local. Use morango UUIDField consistently to avoid duplication.
