+++ kolibri/core/content/models.py
@@ -1,64 +1,28 @@
 """
 These models are used in the databases of content that get imported from Studio.
+Any fields added here (and not in base_models.py) are assumed to be locally
+calculated cached fields. If a field is intended to be imported from a content
+database generated by Studio, it should be added in base_models.py.
 
 
 *DEVELOPER WARNING regarding updates to these models*
 
 If you modify the schema here, it has implications for the content import pipeline
-because other systems also read and write these databases.
-
-Specifically, Kolibri Studio has a modified version of this models.py for generating
-backwards-compatible content databases. Changes made here should be propagated to Studio
-in order to allow for generation of databases with the updated schema.
-
-However, in the case where the modifications are intended only within Kolibri, such
-as fields that provide a local record of expensive to compute properties, these do
-not need to be added to an importable schema version (and will not be used in Studio).
-To allow use of SQLAlchemy updates (for performance reasons), we also maintain an
-updateable copy of the current Django schema.
-
-In order to track updates to models or fields, the CONTENT_SCHEMA_VERSION value must be
-incremented, with an additional constant added for the new version.
-E.g. a new constant VERSION_3 = '3', might be added, and CONTENT_SCHEMA_VERSION set to
-VERSION_3.
-
-In addition, the new constant should be added to the mappings dict in
-./utils/channel_import.py with an appropriate ChannelImport class associated.
-This map associates content schema versions with associated ChannelImport classes,
-which allows proper importing of that content database into the current content schema
-for Kolibri.
-
-If the new schema requires inference of the field when it is missing from old databases
-(i.e. it does not have a default value, or cannot be null or blank), then all the
-ChannelImport classes for previous versions must be updated to infer this data from old
-databases.
-
-A pickled SQLAlchemy schema for the new schema must also be generated using the
-generate_schema management command. This must be generated using an empty, migrated
-database.
-
-The 'version' parameter passed to the command should be the value of e.g. VERSION_3:
-
-    `kolibri manage generate_schema 3`
-
-All pickled schema should be registered in the CONTENT_DB_SCHEMA_VERSIONS list in
-this file e.g. VERSION_3 should be added to the list.
-
-The channel import test classes for the previous schema should also be added in
-./test/test_channel_import.py - it should inherit from the NaiveImportTestCase, and set the
-name property to the previous CONTENT_SCHEMA_VERSION e.g. VERSION_2.
+because we will need to calculate these values during content import (as we they will
+not be present in the content databases distributed by Studio).
 
 In the case where new fields are added that do not need to be added to an export schema
 the generate_schema command should be run like this:
 
     `kolibri manage generate_schema current`
 
-This will just regenerate the current schema and not generate any other schema or fixtures.
+This will just regenerate the current schema for SQLAlchemy, so that we can use SQLAlchemy
+to calculate these fields if needed (this can frequently be more efficient than using the
+Django ORM for these calculations).
 """
 from __future__ import print_function
 
 import os
-import uuid
 from gettext import gettext as _
 
 from django.core.urlresolvers import reverse
@@ -67,92 +31,22 @@ from django.db import models
 from django.db.models import Min
 from django.utils.encoding import python_2_unicode_compatible
 from django.utils.text import get_valid_filename
-from jsonfield import JSONField
 from le_utils.constants import content_kinds
-from le_utils.constants import file_formats
 from le_utils.constants import format_presets
-from le_utils.constants.languages import LANGUAGE_DIRECTIONS
 from mptt.managers import TreeManager
-from mptt.models import MPTTModel
-from mptt.models import TreeForeignKey
 from mptt.querysets import TreeQuerySet
 
 from .utils import paths
+from kolibri.core.content import base_models
 from kolibri.core.content.errors import InvalidStorageFilenameError
 from kolibri.core.device.models import ContentCacheKey
-from kolibri.core.fields import DateTimeTzField
 from kolibri.utils.conf import OPTIONS
 
 PRESET_LOOKUP = dict(format_presets.choices)
 
-V020BETA1 = "v0.2.0-beta1"
-
-V040BETA3 = "v0.4.0-beta3"
-
-NO_VERSION = "unversioned"
-
-VERSION_1 = "1"
-
-VERSION_2 = "2"
-
-# List of the content db schema versions, ordered from most recent to least recent.
-# When a new schema version is generated, it should be added here, at the top of the list.
-CONTENT_DB_SCHEMA_VERSIONS = [VERSION_2, VERSION_1, NO_VERSION, V040BETA3, V020BETA1]
-
-# The latest compatible exported schema version for this version of Kolibri
-CONTENT_SCHEMA_VERSION = VERSION_2
-
-# The version name for the current content schema,
-# which may have schema modifications not present in the export schema
-CURRENT_SCHEMA_VERSION = "current"
-
-
-class UUIDField(models.CharField):
-    """
-    Adaptation of Django's UUIDField, but with 32-char hex representation as Python representation rather than a UUID instance.
-    """
-
-    def __init__(self, *args, **kwargs):
-        kwargs["max_length"] = 32
-        super(UUIDField, self).__init__(*args, **kwargs)
-
-    def prepare_value(self, value):
-        if isinstance(value, uuid.UUID):
-            return value.hex
-        return value
-
-    def deconstruct(self):
-        name, path, args, kwargs = super(UUIDField, self).deconstruct()
-        del kwargs["max_length"]
-        return name, path, args, kwargs
-
-    def get_internal_type(self):
-        return "UUIDField"
-
-    def get_db_prep_value(self, value, connection, prepared=False):
-        if value is None:
-            return None
-        if not isinstance(value, uuid.UUID):
-            try:
-                value = uuid.UUID(value)
-            except AttributeError:
-                raise TypeError(self.error_messages["invalid"] % {"value": value})
-        return value.hex
-
-    def from_db_value(self, value, expression, connection, context):
-        return self.to_python(value)
-
-    def to_python(self, value):
-        if isinstance(value, uuid.UUID):
-            return value.hex
-        return value
-
 
 @python_2_unicode_compatible
-class ContentTag(models.Model):
-    id = UUIDField(primary_key=True)
-    tag_name = models.CharField(max_length=30, blank=True)
-
+class ContentTag(base_models.ContentTag):
     def __str__(self):
         return self.tag_name
 
@@ -189,7 +83,7 @@ class ContentNodeManager(
 
 
 @python_2_unicode_compatible
-class ContentNode(MPTTModel):
+class ContentNode(base_models.ContentNode):
     """
     The primary object type in a content database. Defines the properties that are shared
     across all content types.
@@ -198,40 +92,6 @@ class ContentNode(MPTTModel):
     exist as nodes in content channels.
     """
 
-    id = UUIDField(primary_key=True)
-    parent = TreeForeignKey(
-        "self", null=True, blank=True, related_name="children", db_index=True
-    )
-    license_name = models.CharField(max_length=50, null=True, blank=True)
-    license_description = models.TextField(null=True, blank=True)
-    has_prerequisite = models.ManyToManyField(
-        "self", related_name="prerequisite_for", symmetrical=False, blank=True
-    )
-    related = models.ManyToManyField("self", symmetrical=True, blank=True)
-    tags = models.ManyToManyField(
-        ContentTag, symmetrical=False, related_name="tagged_content", blank=True
-    )
-    title = models.CharField(max_length=200)
-    coach_content = models.BooleanField(default=False)
-
-    # the content_id is used for tracking a user's interaction with a piece of
-    # content, in the face of possibly many copies of that content. When a user
-    # interacts with a piece of content, all substantially similar pieces of
-    # content should be marked as such as well. We track these "substantially
-    # similar" types of content by having them have the same content_id.
-    content_id = UUIDField(db_index=True)
-    channel_id = UUIDField(db_index=True)
-
-    description = models.TextField(blank=True, null=True)
-    sort_order = models.FloatField(blank=True, null=True)
-    license_owner = models.CharField(max_length=200, blank=True)
-    author = models.CharField(max_length=200, blank=True)
-    kind = models.CharField(max_length=200, choices=content_kinds.choices, blank=True)
-    available = models.BooleanField(default=False)
-    stemmed_metaphone = models.CharField(
-        max_length=1800, blank=True
-    )  # for fuzzy search in title and description
-    lang = models.ForeignKey("Language", blank=True, null=True)
     # Fields used only on Kolibri and not imported from a content database
     # Total number of coach only resources for this node
     num_coach_contents = models.IntegerField(default=0, null=True, blank=True)
@@ -261,38 +121,16 @@ class ContentNode(MPTTModel):
 
 
 @python_2_unicode_compatible
-class Language(models.Model):
-    id = models.CharField(max_length=14, primary_key=True)
-    lang_code = models.CharField(max_length=3, db_index=True)
-    lang_subcode = models.CharField(max_length=10, db_index=True, blank=True, null=True)
-    # Localized name
-    lang_name = models.CharField(max_length=100, blank=True, null=True)
-    lang_direction = models.CharField(
-        max_length=3, choices=LANGUAGE_DIRECTIONS, default=LANGUAGE_DIRECTIONS[0][0]
-    )
-
+class Language(base_models.Language):
     def __str__(self):
         return self.lang_name or ""
 
 
-class File(models.Model):
+class File(base_models.File):
     """
     The second to bottom layer of the contentDB schema, defines the basic building brick for content.
     Things it can represent are, for example, mp4, avi, mov, html, css, jpeg, pdf, mp3...
     """
-
-    id = UUIDField(primary_key=True)
-    # The foreign key mapping happens here as many File objects can map onto a single local file
-    local_file = models.ForeignKey("LocalFile", related_name="files")
-    available = models.BooleanField(default=False)
-    contentnode = models.ForeignKey(ContentNode, related_name="files")
-    preset = models.CharField(
-        max_length=150, choices=format_presets.choices, blank=True
-    )
-    lang = models.ForeignKey(Language, blank=True, null=True)
-    supplementary = models.BooleanField(default=False)
-    thumbnail = models.BooleanField(default=False)
-    priority = models.IntegerField(blank=True, null=True, db_index=True)
 
     class Meta:
         ordering = ["priority"]
@@ -355,18 +193,10 @@ class LocalFileManager(models.Manager):
 
 
 @python_2_unicode_compatible
-class LocalFile(models.Model):
+class LocalFile(base_models.LocalFile):
     """
     The bottom layer of the contentDB schema, defines the local state of files on the device storage.
     """
-
-    # ID should be the checksum of the file
-    id = models.CharField(max_length=32, primary_key=True)
-    extension = models.CharField(
-        max_length=40, choices=file_formats.choices, blank=True
-    )
-    available = models.BooleanField(default=False)
-    file_size = models.IntegerField(blank=True, null=True)
 
     objects = LocalFileManager()
 
@@ -393,7 +223,7 @@ class LocalFile(models.Model):
             return None
 
 
-class AssessmentMetaData(models.Model):
+class AssessmentMetaData(base_models.AssessmentMetaData):
     """
     A model to describe additional metadata that characterizes assessment behaviour in Kolibri.
     This model contains additional fields that are only revelant to content nodes that probe a
@@ -401,37 +231,15 @@ class AssessmentMetaData(models.Model):
     ContentNodes with this metadata may also be able to be used within quizzes and exams.
     """
 
-    id = UUIDField(primary_key=True)
-    contentnode = models.ForeignKey(ContentNode, related_name="assessmentmetadata")
-    # A JSON blob containing a serialized list of ids for questions that the assessment can present.
-    assessment_item_ids = JSONField(default=[])
-    # Length of the above assessment_item_ids for a convenience lookup.
-    number_of_assessments = models.IntegerField()
-    # A JSON blob describing the mastery model that is used to set this assessment as mastered.
-    mastery_model = JSONField(default={})
-    # Should the questions listed in assessment_item_ids be presented in a random order?
-    randomize = models.BooleanField(default=False)
-    # Is this assessment compatible with being previewed and answer filled for display in coach reports
-    # and use in summative and formative tests?
-    is_manipulable = models.BooleanField(default=False)
+    pass
 
 
 @python_2_unicode_compatible
-class ChannelMetadata(models.Model):
+class ChannelMetadata(base_models.ChannelMetadata):
     """
     Holds metadata about all existing content databases that exist locally.
     """
 
-    id = UUIDField(primary_key=True)
-    name = models.CharField(max_length=200)
-    description = models.CharField(max_length=400, blank=True)
-    author = models.CharField(max_length=400, blank=True)
-    version = models.IntegerField(default=0)
-    thumbnail = models.TextField(blank=True)
-    last_updated = DateTimeTzField(null=True)
-    # Minimum version of Kolibri that this content database is compatible with
-    min_schema_version = models.CharField(max_length=50)
-    root = models.ForeignKey(ContentNode)
     # precalculated fields during annotation/migration
     published_size = models.BigIntegerField(default=0, null=True, blank=True)
     total_resource_count = models.IntegerField(default=0, null=True, blank=True)
