+++ kolibri/core/content/test/test_upgrade.py
@@ -3,17 +3,16 @@ import uuid
 
 from django.core.management import call_command
 from django.test import TransactionTestCase
+from le_utils.constants import content_kinds
 from mock import call
 from mock import patch
 
-from le_utils.constants import content_kinds
+from .sqlalchemytesting import django_connection_engine
+from kolibri.core.content.constants.schema_versions import CONTENT_SCHEMA_VERSION
 from kolibri.core.content.models import ChannelMetadata
-from kolibri.core.content.models import CONTENT_SCHEMA_VERSION
 from kolibri.core.content.models import ContentNode
 from kolibri.core.content.upgrade import fix_multiple_trees_with_tree_id1
 from kolibri.core.content.upgrade import update_num_coach_contents
-
-from .sqlalchemytesting import django_connection_engine
 
 
 def get_engine(connection_string):
